fix(core): special case of `let`
A workaround for the case when the function is a lambda (normally from a `let` statement).
If we don't do this, the lambda will be synthesized without the argument in the environment.
This may cause a type unification failure when a dependent type in companion with a let-bound variable.
For example:
```
def induction(
    P: ℕ -> 'Type, base: P(ℕ::Zero),
    step: ∀(n: ℕ) -> P(n) -> P(ℕ::Succ(n)), nat: ℕ,
): P(nat) = ...

eval {
    let p = (n: ℕ) => ...
    induction(p, ..., ...)  // <--- here
}
```
will cause:
    "Expected argument type: (p ℕ::Zero), found argument ... with type <some-type>"
but (p ℕ::Zero) is actually identical to <some-type>.